Refactor, refactor, refactor
